+++ Changelog.md
@@ -4,7 +4,7 @@
 
 ## Next release
 
-- LIMIT/OFFSET clause added to SELECT (no standard).
+- OFFSET/FETCH clause added to SELECT (introduced in SQL in 2008).
 - ORDER BY clause added to SELECT (<sort specification>).
 
 ## v0.3.0